making it async

+++ src/domain/rate_limit.rs
@@ -1,5 +1,5 @@
 #[derive(Debug)]
 pub struct RateLimitResult {
     pub allowed: bool,
-    pub remaining: u64,
+    pub remaining: f64,
 }
